+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `Innmind\OperatingSystem\Config::mapSQLConnection()`
 - `Innmind\OperatingSystem\Config::mapServerControl()`
 - `Innmind\OperatingSystem\Config::mapServerStatus()`
+- `Innmind\OperatingSystem\Config::mapClock()`
 
 ### Changed